Fix keepOpen for deleting a child node
